Making levels modular so multiple people can contribute at once.
