+++ mygame/src/dungeons/levels/level2.js
@@ -1,0 +1,31 @@
+const level2 = [
+	"Dungeon - 1",
+	[
+		"XXXXXXXXX",
+		"X        ",
+		"X        ",
+		"X        ",
+		"X        ",
+		"X        ",
+		"X        ",
+		"X        ",
+		"X        ",
+	],
+	[
+		"[========]",
+		"[   @    ]",
+		"[        ]",
+		"[        ]",
+		"[        ]",
+		"[   $    ]",
+		"[        ]",
+		"[        ]",
+		"[        ]",
+		",________."
+	],
+	[
+		"  b    b  "
+	]
+
+];
+export default level2;
